Move `ClientIPFromConn` from `utils/net.go` to `utilsaddr/addr.go`
This function fits better with `addr.go` than the Context specific functions in the rest of `net.go`.  This change will faciliate moving those functions next.
